python/requirements.txt
selenium
webdriver-manager
beautifulsoup4
requests
Pillow
pytesseract
tqdm